[11.2] Premiere version fonctionnelle de ComposeWaitAndRefresh

+++ code/ComposeWaitRefresh/app/src/main/java/com/example/composewaitrefresh/api/Truc.kt
@@ -3,5 +3,5 @@ package com.example.composewaitrefresh.api
 data class Truc (
     val a : Int,
     val b : String,
-    val c : List<Int>
+    val c : List<Double>
 )
